requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.9
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
lxml==6.0.0
numpy==2.3.1
outcome==1.3.0.post0
pandas==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
